Remove the fetcher and initializer.
The dom should be fetched with something else to ensure modularity.
