+++ README.md
@@ -6,4 +6,4 @@ node-goldwasher
 [![Dependency Status](https://david-dm.org/alexlangberg/node-goldwasher.svg)](https://david-dm.org/alexlangberg/node-goldwasher)
 [![devDependency Status](https://david-dm.org/alexlangberg/node-goldwasher/dev-status.svg)](https://david-dm.org/alexlangberg/node-goldwasher#info=devDependencies)
 
-Text scraping/analysis with node.
+Text scraping/cleaning/analysis with node.
