+++ package.json
@@ -37,7 +37,6 @@
     "test": "test"
   },
   "dependencies": {
-    "request": "^2.35.0",
     "cheerio": "^0.16.0",
     "underscore": "^1.6.0",
     "validator": "^3.12.0"
